bugfix
In the simulation I conducted with the AXI GPIO block provided by Xilinx, the READY signals did not come from the AXI GPIO block as long as the AWVALID and WVALID signals were not set to 1. Both the AWVALID and WVALID signals must be ready at the same time, and the address-data information also needs to be ready. Below is the link to Xilinx's explanation:

https://adaptivesupport.amd.com/s/article/63381?language=en_US